refactor: update imports to use new maintenance-request slice file
