+++ client/src/components/partials/home/Index.tsx
@@ -4,7 +4,7 @@ import MaintenanceRequestSummary from '@/components/partials/home/Summary'
 import MaintenanceRequestList from '@/components/partials/home/list/Index'
 import { useEffect } from 'react'
 import { RootState } from '@/lib/store'
-import { summary, findAll } from '@/lib/slices/maintenance-request.slice'
+import { summary, findAll } from '@/lib/slices/maintenance-request'
 import { Button, Link } from '@heroui/react'
 import MaintenanceRequestSubscription from '@/components/partials/home/Subscription'
 import MaintenanceRequestFilter from '@/components/partials/home/Filter'
